Projet-Sensorsbox-serveur: Fix icon font load Models: new type 'float' for number attributes Merge conflicts generated by new editorconfig

+++ Projet-Sensorsbox-serveur/api/models/Record.js
@@ -18,7 +18,7 @@ module.exports = {
 			type: 'float'
 		},
 		timerange: {
-			type: 'int'
+			type: 'float'
 		},
 		sensor: {
 			model: 'sensor'
